Adding automated photometry.
Currently it will try a finely sampled, large range of circular radii with exact and soft edges and trying it with/without moving the aperture.

There should be code after that which finds the best photometry method and then feeds that into the automated decorrelation script. It would be amazing if you just had to download the zip files from SHA and then click go on the pipeline, and it finished everything after that for you.