Increasing automation.
Automated the creation of the required directory structure. The only thing needed of the user at this point is downloading the zip files from the Spitzer Heritage Archive, and telling the code where to find those files. It then unzips them, puts all the files in the right place, and makes all the necessary folders.

Also fully automated the photometry step of the reduction. It will now try a wide range of different apertures, with/without a dark correction stack, with/without ignoring bad frames which it finds automatically, and with/without a moving aperture. It will then decide which photometry performs the best and write that to a file that can later be read by the decorrelation notebook once I have that coded.