Resolving issue #11.
There was a bug where x and y coordinates were swapped in the centroid_FWM function and the oversampling function due to using numpy.meshgrid, as was noticed by Dylan. This has now been patched, and unit tests have been written to make sure this doesn't happen in the future.