Making george optional.
Only required if the user wants to run GPs. It'll just fully crash if they try to run a GP without george though. Should make this a softer crash in the future. Also, message about george everytime you import SPCA will be pretty annoying for those that don't want to use it. Improve this in the future.