Adding debugged PLD decorrelation.
Seemed to run from start to finish fine for MASCARA-1b data with 3x3 2nd order PLD. Other burn-in attempts suggest that PLD1 and 5x5 are all fine too. The fit was quite poor for MASCARA-1b though. I'll run it overnight on the other planets and see if it does any better, but perhaps there's an underlying issue.