Trying removing threadpool_limits.
There is a known bug where (only for some data sets and only for some decorrelation methods for those data sets), the emcee runs will inexplicably hang part way through. This is moderately repeatable (will almost always keep happening for the same dataset+decorrelation combination, but not always after the same amount of steps or runtime), and Ian Crossfield has also encountered this (see Issue #16). What happens is that all but one thread will suddenly stop working, and the last remaining thread will run at 100 percent for as long as you let it go. I never encountered this when using single threaded emcee, so I suspect this has something to do with the interaction between emcee and multiprocessing.pool and/or threadpoolctl.threadpool_limits. This will be an especially hard bug to sort out as it is difficult to debug emcee run, let along a multi-threaded emcee run.

As a first step, I'm getting rid of the use of threadpoolctl.threadpool_limits as some local tests on my laptop seem to suggest it actually doesn't appreciably speed things up. I'll try to find the dataset+decorrelation combos that froze for me before and see if they still freeze (I've asked Ian to do the same on the master branch). If so, then at least the problem has been narrowed down somewhat. If not, then problem solved (without really understanding the root of the problem). emcee really seems to use multiprocessing.pool very simply, so I have a hard time believing that interface could be the issue though.