New code to correctly compute photon noise limit.
Adding new code to correctly compute the photon noise limit either for old photometry that was saved with bad units or for new data saved in units of electrons. Eventually we will remove compatibility with old photometry, but for now its much easier to have this backwards compatibility rather than re-running all of the photometry for the big re-analysis of previously published phasecurves.