Initial workaround for MCMC hanging.
Still not sure what is causing the MCMC runs to freeze spontaneously (see Issue #16), but the general issue is that at least one of the walkers freezes and the whole process has to wait indefinitely for that one walker to return. By using the open source package timeout_decorator, I am able to set a 10 second maximum execution time on the log-likelihood function, catch the TimeoutError that gets thrown if the time limit is exceeded, print out the walker position, and return -np.inf. At the least this will allow the MCMC to resume, and hopefully the printed p0 will help me understand the state when the function freezes and debug the underlying issue.