Speeding up aperture, allow for large datasets.
Speeding up aperture photometry but using multiprocessing in a better place and by avoiding duplications wherever possible. Allowing for large (>4GB) datasets with multiprocessing by using a single global variable.

This current commit is part way through the process of updating the code, and a bunch of things are likely going to be broken. The biggest cause of current bugs is likely going to be comparing RMSs for different photometries. I'll come back to this tomorrow and get it back and running.