Improving the sigma clipping of photometry data.
Adding a step where arrays are sigma clipped at 10 sigma, temporarily smoothed using a box-car filter, and then sigma clipped at 5 sigma. This helps to remove moderately bad data points. Also, instead of just discarding the data, it is replaced with the median of the two points to its left and the two points to its right.