Updates related to BLISS knot selection.
The utilized number of BLISS knots are now saved and can be reloaded with the initializeWithOld option or just runMCMC=False. Also making a correction to the number of utilized knots where before it was a sum over the NNI array, it is now a sum over the BLS array from bliss.precompute().

Also fixing minor bug in Decorrelation.py which displayed plotted the photometry.