Fixing bug with CPU-bound sigma clipping.
Previously, it was possible for large swaths of the data to get sigma clipped using the new CPU-heavy sigma clipping method rather than the RAM-heavy method used earlier. This is fixed by flattening the arrays and removing the axis arg (the latter seemed to be the issue). Also, the mask for ignoreFrames was not being properly set before, but is now.

Also turning 5x5 PLD back on by default for Everything_Photometry.ipynb.